requirements.txt
django
djangorestframework
gunicorn==20.1
python-dotenv
pillow